feat: added login (jwt token, hashed password) and oauth, foreign key connections in the gorm entities

+++ entity/user.go
@@ -2,7 +2,7 @@ package entity
 
 // Entities are the results we'd expect from the database.
 type User struct {
-	UserID       int    `gorm:"column:id;primary_key"`
+	UserID       int64  `gorm:"column:id;primary_key"`
 	Name         string `gorm:"column:name"`
 	Email        string `gorm:"column:email"`
 	Password     string `gorm:"column:password"`
